back-end/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.32.0
pydantic==2.9.2
openai==1.59.5
google-genai==2.0.1
python-dotenv==1.0.1
qdrant-client==1.12.1
pypdf==5.1.0
python-multipart==0.0.12
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.1
